+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -18,10 +18,13 @@ exclude: true
 - Added [`CompositorController.onPipModeChanged`][139.1] to [`CompositorController`][65.1] to inform GeckoSession of changes to and from picture-in-picture mode.
 - ⚠️Increased `compileSdkVersion` to 36 (Android 16)
 - Added [`getLowMemoryDetection()`][139.2] on `GeckoRuntimeSettings` and [`lowMemoryDetection(boolean)`][139.3] on `GeckoRuntimeSettings.Builder` to control the low-memory detection machinery.
+- Added `setEnhancedTrackingProtectionCategory` to [`ContentBlocking.Settings`][139.4].
+  ([bug 1956620]({{bugzilla}}1956620))
 
 [139.1]: {{javadoc_uri}}/CompositorController.html#onPipModeChanged
 [139.2]: {{javadoc_uri}}/GeckoRuntimeSettings.html#getLowMemoryDetection
 [139.3]: {{javadoc_uri}}/GeckoRuntimeSettings.Builder.html#lowMemoryDetection
+[139.4]: {{javadoc_uri}}/ContentBlocking.html
 
 ## v138
 - Added [`GeckoSession.Loader.originalInput`][138.1] option, which allows passing through the original user address bar input
@@ -1709,4 +1712,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 3bd4ce6a972868cd2c50cce9047b0b4cb87376b7
+[api-version]: bba475b7cb5de5ca051e20dad99084d132ebc9d3